feat : frontend RepositoryActivitySection 컴포넌트 분리
